Rename files for better practice
